requirements.txt
cozepy
python-dotenv
fastapi[standard]
dashscope
pymysql
sqlalchemy
aiofiles
oss2
pytest
pytest-asyncio
ffmpeg-python
volcengine-python-sdk[ark]
